refactor(observability): 分离体验时间线投影
